+++ lib/rszr.rb
@@ -3,13 +3,14 @@ require 'pathname'
 require 'tempfile'
 require 'stringio'
 
-require 'rszr/rszr'
 require 'rszr/version'
 require 'rszr/stream'
 require 'rszr/identification'
 require 'rszr/orientation'
 require 'rszr/buffered'
 require 'rszr/color'
+require 'rszr/fill'
+require 'rszr/rszr'
 require 'rszr/image'
 
 module Rszr